docgenerator/requirements.txt
Flask==3.0.0
python-dotenv==1.0.1
gunicorn==22.0.0

google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-oauthlib==1.2.1
gspread==6.2.1
oauth2client==4.1.3

requests==2.31.0
email-validator==2.1.1